Upload: move S3Agent out cli/upload_command.py
